Only update registered indexes when calling updateAllIndexes
Previously, IndexManager would try to update all indexes that exist
in the index database. This would cause an exception for those which
hadn't been registered this session with ensureIndexed as the
indexing function for those indexes wasn't known.

This patch just excludes unregistered indexes from the update process
and adds a regression test.