Work-around for base64 issues:
Android uses an old version of commons-codec (1.2?) which doesn't have
Base64{Input,Output}Stream. Their version always takes priority so we can't use a newer version. So,
if running on android, use the android.util libraries for this purpose.

NB once the 31028-java-desktop-build branch is moved, the classes will need to be loaded by
class.forName() to de-couple them.