Run tests on android
Created an new project within sync-android which is a application designed to run JUnit 4 tests, with all the benefits. The test application has several
flavours which can be used to test sync-android, standardTest, performanceTest, integrationTest, systemTest. These flavours represent test types which can be
performed using the gradle tasks in sync-core. The test application requires fixtures to be uploaded using the uploadFixtures task, this is a dependancy of startTestApp.
waitForTestAppToFinishTests task has a depdancy on startTestApp, and will ask the device/emulator ( at approx 1 minute intervals ) if a test results file is present
on the sdcard ( /sdcard ), when the file is dectected and downlaoded, the task will finish executing.

Fixture loading has been changed. Fixtures now need to be loaded via TestUtils.loadFixture(Sting filename) method. This method will create a file object that points
to the correct place on disk. For android this is in the sdcard dir, for java se it is in the working directory for the JVM.

BuildConfig class is used to pass in the test classes to run, and specifiy which classes should be excluded using JUnit categories.