Reduce use of internal numeric id
This commit does several things, firstly it removes methods which expose
the internal numeric id for a document from the public. These methods
are:

- DatastoreExtended#getDocNumericId()
- BasicDocumentRevision#getInternalNumericId()

Secondly it removes the method
BasicDatastore#getAllRevisionsOfDocument(long docNumericID) this is
beacause its funtionalty could easily be moved into the
BasicDatastore#getAllRevisionsOfDocument(String documentId) with a small
modification to the SQLQuery.

Thirdly it removes BasicDatastore#getDocumentNumericId replacing its
function by adding the documentNumericId property to DocumentRevisionTree

Finally in order to make the DocumentRevisionTreeTest more readable, the
document revisions are renamed and the correct internalDocumentID is
assigned to all instances created in the set up.