Add single level AND support to query handler
- Simple query can execute from IndexManager find(query) method.  Query
can be a single equality or a single level ANDing of multiple
equalities.  For example query can be { "name" : "mike" } or it can be
{ "name" : "mike", "age" : 12 }.
- At present query find works only if valid index(es) exist for the
defined query.
- Tests included to support functionality detailed in above two bullets.
- Remove QueryResult interface and rename BasicQueryResult class as
QueryResult.  Going forward QueryResult will be a class.
- For code clarity, replace any reference to .iterator().next() with
.toArray()[0] when no actual iteration is being performed.
- Added tests to QuerySqlTranslatorTest that test when selecting an
index to use, when generating query where clauses, and when generating
query select clauses.
- Add comment to TranslateState class clarifying its purpose and
implementation.
- Fix class level comments throughout index/query.
- Change evaluation of top level query node in QuerySqlTranslator for
to use Map.get method.  $and/$or now compared to != null as success
condition.
- Fix "not" operator check to check for "$not".
- Change fieldName definition from ArrayList to List in
IndexManager.ensureIndexed and throughout in order to make method more
dynamic.
- For code clarity, in test classes change the population of lists to
use Arrays.asList instead of .add method of the List itself when all
values of the List are already known.
- Move QueryResults iterator test into its own test method inside of
QueryExecutorTest.  Remove repeated tests of same functionality in
other QueryExecutorTest methods.
- Fix bug in QueryExecutor.executeQueryTree.  The accumulator of doc ids
is to accumulate by set intersection.
- Add SqlParts class which provides a way to store SQL with placeholders
and the values for those placeholders separately.
- Change the nature of the SQL generation process for query retrieval
from String replacement to using the SqlParts class with placeholders
and values.
- Update class level comments for QueryNode, ChildrenQueryNode,
AndQueryNode, OrQueryNode
- Remove the new ArrayList<Object>(Arrays.asList(..)) definition from
fieldName references in tests and replace with Arrays.<Object>asList(..)
- Remove the now unecessary supression of unchecked casting warning
where appropriate.