Execute all datastore db operations on a serial queue.
All database operations are performed on a single serial queue,
this stops leaking of connections when a user's application uses the
datastore on different threads, but only calls close on a single
thread.

Transactions can be handled using the serial queue SQLDatabseQueue by
calling submitTransaction method. When a callable is called in a
transaction throwing an exception causes the transaction to be rolled
back.

forceInsert method on DatastoreExtended interface now takes an
additional parameter, preparedAttachments, which is contains attachments
that were download seperately to the document. This removes the need for
the pull replicator to start transactions on Datastore object.

Remove the getSQLDatabase from the DatastoreExtended interface

Remove threadlocal code from SQLiteWrapper, enforcing the need to perform
each database operation on the same thread.

Added a TestUtil method to create a connection to underlying database
for the datastore.

Remove test which checks if the underlying db is closed when
datastore.close() is called.