Correct encoding of database URLs:
Some refactoring to ensure the user passes in a correctly encoded URL
for remote databases. This includes removing the 'default DB'
anti-pattern in the Mazha library.

This way, we don't try and second-guess the path structure which
breaks reverse proxies/forwarding etc.

Usernames and passwords have been removed from a number of classes as
these are now assumed to be part of the 'root' URL. This and the
removal of 'default DB' has resulted in a large refactor of a number
of test classes.