Use checked exceptions for recoverable errors
This commit changes how exceptions are handled in sync-android.
Previously, virtually all exceptions were unchecked. While
convenient, this made it too easy to ignore fairly common
occurances like conflict exceptions when saving documents. We
decided it was worth exposing the recoverable failure modes
as checked exceptions. An outline of the exceptions are below.

- DatastoreException
  |
  \- DatastoreNotCreatedException

DatastoreExceptions are for failures of the datastore, and
usually indicate fairly exceptional conditions -- SQL
execution failures, disk space issues and so on.

- DocumentException
  |
  |- InvalidDocumentException
  |- DocumentNotFoundException
  |- ConflictException
  |
  \- AttachmentException
        |
        \- AttachmentNotSavedException

DocumentExceptions usually indicate issues with specific
document operations -- invalid JSON, conflict on update,
nonexistent documents and the like. These states are
usually recoverable in code (e.g., get the latest doc
revision to fix a ConflictException).

AttachmentExceptions are usually thrown for problems
where the attachment cannot be retrieved, for example
a document which is being pulled from a HTTP stream.