Add SQLCipher-based encryption for main JSON data
This commit adds the KeyProvider interface which holds an
encryption key. The provider is passed through the datastore
opening methods down to the database layer, where a new class
AndroidSQLCipherSQLite opens the database using the provided
key and SQLCipher rather than using the existing
AndroidSQLite class. AndroidSQLite is still used for datastores
which are not encrypted.

We use only 256-bit AES keys, which we pass to SQLCipher as
raw key material to avoid SQLCipher using PBKDF2 to derive
a key. This is because we need a secure key for encrypting
attachments in future commits, so we might as well get
SQLCipher to use the same key.

For testing, a SimpleKeyProvider is implemented, which just
takes a byte[] key as input and supplies it when asked.

There is also a test.sqlcipher.passphrase define added which
makes DatastoreTestBase use SQLCipher during testing.

The commit also adds required binaries (jar and .so files) for
SQLCipher testing to the sample application.