Use raw 32-byte key for SQLCipher rather than password
This commit changes how we initialise SQLCipher to use a 32-byte
key rather than a password, which SQLCipher would then use to derive
a key using PBKDF2. As we need to manage a secure key for attachments,
we'd prefer SQLCipher to use the same key and not waste time deriving
a key from our key.

In addition to changing what we provide SQLCipher, this commit also
adds classes which validate the key is the correct length for passing
to SQLCipher along with utility functions to create appropriate
key strings for SQLCipher.