Change database migration to always use a Migration object
This unifies the call path for when one just has SQL to run and
when one needs to run Java code as part of a migration.

I added a helper class, RunSchemaOnlyMigration for the case
when there is only SQL to run.

Using an explicit migration inteface makes it clear that a
class is a migration, rather than generic SQL callable code.