Add EndToEndEncryptionTest to test files encrypted on disk
This test creates a database with a key using the DatastoreManager.
It then looks at the JSON database, the indexes database and the
attachment file created to check they are encrypted or not, using
parameterisation to check both encrypted and non-encrypted
datastores.

End to end tests are in sync-android project so they are only
tested on Android as JavaSE doesn't support encryption yet.

It also parameterises the test so we _always_ test both the
encrypted and non-encrypted case when running on Android for
the EndToEndEncryptionTest case.

I copied TestUtils into the sync-android library to avoid
dependency issues with the non-Android build where the class
wasn't available to the sync-android sub-project compilation.