Don't use encryption in testing based on envvar
This commit removes the test.sqlcipher.passphrase env var
which enabled encryption on all tests descending from
DatastoreTestBase. It was causing issues for running the
tests on JavaSE and made it difficult to run encryption testing
in IDEA/Studio.

Instead, we'll have a specific set of tests which run on Android
which run though some encryption basics. Ideally we'd like to run
all tests both with and without encryption, but we can move to
that later.