Added task cancellation when replication stopped
Added a list of tasks in BasicPullStrategy.
Added a cancel method on GetRevisionTaskThreaded.
Tracked submitted jobs in GetRevisionTaskThreaded and cancel them and
drain down the completed response queue when cancel is called.

Added new test to assert tasks are removed when replication stopped.

Added CHANGES.md entry for this fix