Filter Push Replications
Made it possible to filter push replications. In order to this a user needs to implement
the PushFilter interface. The interface method is called during changes feed generation, this filters
the changes feed before being passed to the replicator.

Put filtering and checkpointing code in the `replicate()` method.
Stored the unfiltered size and last seq.
Made checkpointing based on if there were any changes,
not just changes that were processed.

Added tests for filtered push replications.

Updated CHANGES.md.

Fixes #214