Fixed pickWinner for multiple roots with same rev
In the case of multiple roots with the same rev getting the sequence
number by using the revid is not guaranteed to be correct. Instead
changed to collect the sequence ID at the same time as the revid and
used a SortedMap to store both.

Added test for picking winner of two identical root nodes.

Updated CHANGES.md.