Move most SQL code out of DatastoreImpl
This is achieved by moving the SQL code out of the various `*InQueue`
methods or methods in anonymous classes. Instead each 'operation'
becomes a `*Callable` class which overrides the `call()` method.

This change widens access of some fields and methods from
package-private to public, due to the fact that callables live in their
own package. This may not be entirely desirable and other work-arounds
could be considered.