Point users to latest code
For the sample application some users download the current
version on master, this at times relies on code that has yet
to be released. To combat this a small note (in bold) has
been added to the README pointing to the latest release page,
where the code can be downloaded.