Receive the latest revision for a tree.
When calling open_revs, it is possible that the
revision that was retrieved from the changes feed
may not be available eg has been deleted. To
combat this, adding the `latest=true` to the query
means that the server will return an ancestor
to the rev requested even if the revision has
been deleted.

There is a problem however, since the server may
still return a missing revision in the event that
the revision is truly missing, the replicator will
still error in the event that a missing rev is encountered.