Remove batch limit for replications.
Remove the limit of the number of batches that can run for replications,
this ensures that a replication will replicate until there are no more
changes to be replicated from the database.

If a running replication is taking a long time to complete, the
replication can be ended by canceling the replication via the
replicator.

Resolves #462