Ensure attachments are ordered
Copy attachments from DocumentRevision to a TreeMap to ensure an
ordering of attachment entries (they are ordered by the key which is the
attachment name).

This ensures that multipart bodies are output in the same order as their
entries in the _attachments map in the json.

Add test attachmentOrderingTest which shuffles 1000 attachments and
ensures that the output is in the correct order.