Re-write MultipartAttachmentWriter
The previous implementation used a potentially large number of input
streams which caused excessive file handles to be open
simultaneously. This implementation uses a state machine to track the
current part being written and opens and closes streams on demand, so
that there is only ever one stream open at a time.